+++ lib/getErrorsAndWarnings.js
@@ -1,3 +1,10 @@
+/**
+ * @method getErrorsAndWarnings Given the information update errors and warnings array in resultInfo
+ * @param {object} assertionObject Provides the assertion and the results of the test
+ * @param {object} assertionInfo Information about the assertion
+ * @param {object} resultInfo Arrays to store error and warning info
+ */
+
 function getErrorsAndWarnings(assertionObject, assertionInfo, resultInfo) {
     const { type, assertionId, test, simplifiedTest, description, results } = assertionObject;
     const { patternId, ruleId, context } = assertionInfo;
