+++ lib/getFailedAssertions.js
@@ -1,5 +1,14 @@
 const _get = require('lodash.get');
-const checkRule = require('./checkRule');
+const { checkRule } = require('./checkRule');
+
+/**
+ * @method getFailedAssertions Get assertions and extension from the rule object and check the rule against the context
+ * @param {string} ruleId Rule identifier
+ * @param {object} assertionInfo Object to store assertion info
+ * @param {object} options Options passed to the validator and any other variables passed between the methods
+ * @returns {array} array of failedAssertionObjects
+ */
+
 function getFailedAssertions(ruleId, assertionInfo, options) {
     let failedAssertions;
     assertionInfo.ruleId = ruleId;
