+++ lib/getSchematronMap.js
@@ -5,7 +5,7 @@ const parseSchematron = require('../parseSchematron');
 /**
  * @method getSchematronMap A method to get the schematron map from the provided schematron or utilize the data from cache or options object
  * @param {string} schematron String data of schematron
- * @param {object} options options object passed to `validate` or `validateAsync` method
+ * @param {object} options Options passed to the validator and any other variables passed between the methods
  * @returns {object} parsed schemtaron map object
  */
