+++ lib/initializeProcess.js
@@ -9,13 +9,8 @@ const getSchematronMap = require('./getSchematronMap');
  * @param {string} xml string content of an xml file or path to an xml file
  * @param {string} schematron string content of a schematron file or path to a schematron file
  * @param {object} options object that takes parameters as seen in the `validate` method
- * @returns Sets the following variables with file scope
- * xml, schematron, xmlDoc,
- * schematronMap, namespaceMap,
- * patternRuleMap, ruleAssertionMap,
- * resourceDir, xmlSnippetMaxLength,
- * xpathSelect, errors, warnings,
- * ignored, and resultInfo
+ * Sets some method specific variables that will be passed to other methods using the options object
+ * options.xmlDoc, options.schematronMap, options.resourceDir, options.xmlSnippetMaxLength, and options.xpathSelect
  */
 
 function initializeProcess(xml, schematron, options) {
