+++ lib/processRules.js
@@ -3,7 +3,7 @@ const processFailedAssertions = require('./processFailedAssertions');
 /**
  * @method processRules Entry method to start processing schematron rules. Calls methods `getFailedAssertions` and `processFailedAssertions`
  * @param {Array} rules Array of ruleIds
- * @param {object} resultInfo Capturing the information of the rule to pass it on to reporting object
+ * @param {object} resultInfo Object that hosts the errors, warnings and ignored arrays to store the results
  * @param {object} options Passing on the options object passed to the `validate` or `validateAsync` methods
  */
